Database (receiver) now sends responses. Added new requests and request content handlers
