--- client/src/main/java/exceptions/IllegalManagerArgumentException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class IllegalManagerArgumentException extends Exception {
-    public IllegalManagerArgumentException(String message) {
-        super(message);
-    }
-}
