+++ common/src/main/java/commands/CommandDescription.java
@@ -4,7 +4,6 @@ import java.io.Serializable;
 
 public class CommandDescription implements Serializable {
     private final String name;
-    private String[] args;
 
     public CommandDescription(String commandName) {
         this.name = commandName;
@@ -12,13 +11,5 @@ public class CommandDescription implements Serializable {
 
     public String getCommandName() {
         return name;
-    }
-
-    public String[] getArgs() {
-        return args;
-    }
-
-    public void setArgs(String[] args) {
-        this.args = args;
     }
 }
