+++ common/src/main/java/requests/ExitCommandExecutionRequest.java
@@ -1,0 +1,4 @@
+package requests;
+
+public class ExitCommandExecutionRequest implements Request {
+}
