+++ server/src/main/java/commandsModule/commands/ClearCommand.java
@@ -1,7 +1,7 @@
 package commandsModule.commands;
 
 import commands.CommandDescription;
-import dataBase.Database;
+import database.Database;
 
 public class ClearCommand implements BaseCommand {
     private final String name = "clear";
