+++ server/src/main/java/commandsModule/commands/RemoveByIdCommand.java
@@ -1,7 +1,7 @@
 package commandsModule.commands;
 
 import commands.CommandDescription;
-import dataBase.Database;
+import database.Database;
 
 import java.io.IOException;
