+++ server/src/main/java/main/Server.java
@@ -1,5 +1,6 @@
 package main;
 
+import database.Database;
 import requests.Request;
 import serverModules.connection.ConnectionModuleConfigurator;
 import serverModules.connection.ConnectionModule;
@@ -15,12 +16,14 @@ public class Server {
     public static void main(String[] args) {
 
         ConnectionModule module = new ConnectionModuleConfigurator().newInstance(PORT);
+        Database database = new Database();
 
         try {
             while (true) {
                 RequestData requestData = module.receiveData();
                 Request request = new RequestReader().readRequest(requestData.getByteArray());
                 ServerContext context = new ServerContext(module, requestData.getCallerBack(), request);
+                database.setServerContext(context);
                 new RequestHandlerManager().manageRequest(context);
             }
         } catch (Exception e) {
