+++ server/src/main/java/serverModules/context/ServerContextContainAble.java
@@ -1,0 +1,15 @@
+package serverModules.context;
+
+import requests.Request;
+import serverModules.callerBack.CallerBack;
+import serverModules.connection.ConnectionModule;
+
+public interface ServerContextContainAble {
+
+    ConnectionModule getConnectionModule();
+
+    CallerBack getCallerBack();
+
+    Request getRequest();
+
+}
