+++ server/src/main/java/serverModules/request/contentHandlers/ExitCommandExecutionRCH.java
@@ -1,0 +1,11 @@
+package serverModules.request.contentHandlers;
+
+import serverModules.context.ServerContext;
+
+public class ExitCommandExecutionRCH implements RequestContentHandleAble {
+
+    @Override
+    public void handleRequestContent(ServerContext context) {
+        // cooooode
+    }
+}
